+++ tests/src/check_set_details.cpp
@@ -16,8 +16,8 @@ TEST(CheckSetDetails, Success) {
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
 
-    std::vector<int32_t> ranges { 31, 27 };
-    std::vector<int32_t> sizes { 51, 82 };
+    std::vector<uint64_t> ranges { 31, 27 };
+    std::vector<uint64_t> sizes { 51, 82 };
     gesel::internal::check_set_details(path, ranges, sizes);
 }
 
@@ -30,7 +30,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 51, 82 }); }, "tab");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 51, 82 }); }, "tab");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -38,7 +38,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 51, 82 }); }, "newline");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 51, 82 }); }, "newline");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -46,7 +46,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31 }, std::vector<int32_t>{ 51 }); }, "number of lines");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31 }, std::vector<uint64_t>{ 51 }); }, "number of lines");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -54,7 +54,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 28 }, std::vector<int32_t>{ 51, 82 }); }, "bytes per line");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 28 }, std::vector<uint64_t>{ 51, 82 }); }, "bytes per line");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -62,7 +62,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 51, 82 }); }, "early termination");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 51, 82 }); }, "early termination");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -70,7 +70,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nfoo bar\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 51, 82 }); }, "different name");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 51, 82 }); }, "different name");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -78,7 +78,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tfoobar\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 51, 82 }); }, "different description");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 51, 82 }); }, "different description");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -86,7 +86,7 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27 }, std::vector<int32_t>{ 52, 82 }); }, "different size");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27 }, std::vector<uint64_t>{ 52, 82 }); }, "different size");
 
     {
         byteme::RawFileWriter rwriter(path);
@@ -94,5 +94,5 @@ TEST(CheckSetDetails, Failure) {
         byteme::GzipFileWriter gwriter(path + ".gz");
         gwriter.write("aaron's set\tthis is aaron's set\t51\nanother set\tyet another set\t82\n");
     }
-    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<int32_t>{ 31, 27, 42 }, std::vector<int32_t>{ 51, 82, 93 }); }, "number of lines");
+    expect_error([&]() { gesel::internal::check_set_details(path, std::vector<uint64_t>{ 31, 27, 42 }, std::vector<uint64_t>{ 51, 82, 93 }); }, "number of lines");
 }
